third_party/WebKit/LayoutTests/fast/events/scroll-event-handler-count-expected.txt

This test checks that we correctly update the scroll event handler count as event handlers are added and removed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.internals.scrollEventHandlerCount(document) is 0
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS window.internals.scrollEventHandlerCount(document) is 2
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS window.internals.scrollEventHandlerCount(document) is 0
PASS window.internals.scrollEventHandlerCount(document) is 0
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS window.internals.scrollEventHandlerCount(document) is 1
PASS window.internals.scrollEventHandlerCount(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
